+++ src/Menu.py
@@ -1,12 +1,8 @@
-# ===================== import ===================== #
-
 import pygame
 from PIL import Image
 from Game import Game
 from Input import KeyInput
 import os
-
-# ===================== menu ===================== #
 
 class Menu():
     """
@@ -15,15 +11,12 @@ class Menu():
 
     def __init__(self):
         pygame.display.init()
-        self.__keyInput = KeyInput()
         self.__screenSizeInfo = pygame.display.Info()
         self.__screenResolution = pygame.math.Vector2(self.__screenSizeInfo.current_w, self.__screenSizeInfo.current_h)
         self.__screen = pygame.display.set_mode((self.__screenResolution.x, self.__screenResolution.y))
+
         self.__clock = pygame.time.Clock()
-        self.__mousePosition = pygame.mouse.get_pos()
-        self.__currentLevel = None
-        self.__currentLevelBackground = None
-        self.__currentCharacterSkin = None
+        self.__clock_tick = 30
 
     def drawButton(self, buttonImagePath :str, buttonOffsetX :int, buttonOffsetY :int):
         """
@@ -50,148 +43,26 @@ class Menu():
         # returns the coordinates of the rect to test for interactions
         return buttonRect
     
-    def drawBackground(self, __backgroundImagePath :str):
+    def drawBackground(self, backgroundImagePath:str):
         """
         Method to blit a background image onto the screen with a given path.
         """
 
         # uses the given path to draw the background
-        __background = pygame.image.load(__backgroundImagePath)
-        __background = pygame.transform.scale(__background,(self.__screenResolution))
-        self.__screen.blit(__background, (0, 0))
+        background = pygame.image.load(backgroundImagePath)
+        background = pygame.transform.scale(background,(self.__screenResolution))
+        self.__screen.blit(background, (0, 0))
 
-    def mergeImage(self, __imageTopPath :str, __imageBottomPath :str, __newImageName :str):
+    def mergeImage(self, imageTopPath :str, imageBottomPath :str, newImageName :str):
         """
         Method to merge two images to create a single image
         """
         
-        __imageTop = Image.open(__imageTopPath)
-        __imageBottom = Image.open(__imageBottomPath)
-        __imageBottom.paste(__imageTop, (0,0))#, mask = __imageTop)
-        __imageBottom.save(__newImageName) 
-
-# ===================== main menu ===================== #
-
-    def drawMainMenu(self):
-        """
-        Method to fully initate the creation of the main menu.
-        """
-
-        pygame.display.set_caption("super main menu")
-
-        # loads in the objects and draws the main menu
-        self.drawBackground("sprites/placeholder/mainmenu.png")
-        self.__buttonPlayRect = self.drawButton("sprites/placeholder/buttonplay.png", 150, 0)
-        self.__buttonQuitRect = self.drawButton("sprites/placeholder/buttonquit.png", -150, 0)
-        pygame.display.flip()
-
-    def mainMenuLoop(self):
-        """
-        Loop used for the main menu. Returns "levelmenu" or "quit" on specific input, otherwise returns "mainmenu".
-        """
-
-        # frame and input update
-        self.drawMainMenu()
-        self.__keyInput.getInput()
-        self.__mousePosition = pygame.mouse.get_pos()
-
-        # input check
-        if self.__keyInput.keymouseleft:
-
-            if self.__buttonPlayRect.collidepoint(self.__mousePosition):
-                self.__keyInput.keymouseleft = False
-                return "levelmenu"
-        
-            if self.__buttonQuitRect.collidepoint(self.__mousePosition):
-                return "quit"
-            
-        if self.__keyInput.keyescape: # thats the RAGEQUIT button :D
-            return "quit"
-
-        # no input --> reinitialises own loop
-        self.__clock.tick(30)
-        return "mainmenu"
-
-# ===================== level menu ===================== #
-
-    def drawLevelMenu(self):
-        """
-        Method to fully initate the creation of the level menu.
-        """
-
-        pygame.display.set_caption("super level menu")
-
-        # loads in the objects and draws the level menu
-        self.drawBackground("sprites/placeholder/levelmenu.png")
-        self.__level_1_Rect = self.drawButton("sprites/icons/level_icons/level1.png", 250, 500)
-        self.__level_2_Rect = self.drawButton("sprites/icons/level_icons/level2.png", 250, 250)
-        self.__level_3_Rect = self.drawButton("sprites/icons/level_icons/level3.png", 250, 0)
-        self.__level_4_Rect = self.drawButton("sprites/icons/level_icons/level4.png", 250, -250)
-        self.__level_5_Rect = self.drawButton("sprites/icons/level_icons/level5.png", 250, -500)
-
-        pygame.display.flip()
-
-    def levelMenuLoop(self):
-        """
-        Loop used for the level menu. Returns "levelmenu" or "mainmenu" on specific input, otherwise returns "levelmenu".
-        """
-
-        # frame and input update
-        self.drawLevelMenu()
-        self.__keyInput.getInput()
-        self.__mousePosition = pygame.mouse.get_pos()
-        self.__tiles_path = []
-
-        # input check
-        if self.__keyInput.keyescape:
-            self.__keyInput.keyescape = False
-            return "mainmenu"
-
-        if self.__keyInput.keymouseleft and self.__level_1_Rect.collidepoint(self.__mousePosition):
-            self.__currentLevel = "sprites/blocks/csv/level1_grassland.csv"
-            self.__currentLevelBackground = "sprites/placeholder/level1background.png"
-            self.__keyInput.keymouseleft = False
-            self.__tiles_path = self.get_file_names("sprites/blocks/grassland")
-            print(self.get_file_names("sprites/blocks/grassland"))
-            return "charactermenu"
-        
-        if self.__keyInput.keymouseleft and self.__level_2_Rect.collidepoint(self.__mousePosition):
-            self.__currentLevel = "sprites/blocks/csv/level2_desert.csv"
-            self.__currentLevelBackground = "sprites/placeholder/level1background.png"
-            self.__keyInput.keymouseleft = False
-            self.__tiles_path = self.get_file_names("sprites/blocks/desert")
-            print(self.__tiles_path)
-            return "charactermenu"
-        
-        if self.__keyInput.keymouseleft and self.__level_3_Rect.collidepoint(self.__mousePosition):
-            self.__currentLevel = "sprites/blocks/csv/level3_grassland.csv"
-            self.__currentLevelBackground = "sprites/placeholder/level1background.png"
-            self.__keyInput.keymouseleft = False
-            self.__tiles_path = self.get_file_names("sprites/blocks/grassland_2")
-            print(self.__tiles_path)
-            return "charactermenu"
-        
-        if self.__keyInput.keymouseleft and self.__level_4_Rect.collidepoint(self.__mousePosition):
-            self.__currentLevel = "sprites/blocks/csv/level4_snowland.csv"
-            self.__currentLevelBackground = "sprites/placeholder/level1background.png"
-            self.__keyInput.keymouseleft = False
-            self.__tiles_path = self.get_file_names("sprites/blocks/snowland")
-            print(self.__tiles_path)
-            return "charactermenu"
-        
-        if self.__keyInput.keymouseleft and self.__level_5_Rect.collidepoint(self.__mousePosition):
-            self.__currentLevel = "sprites/blocks/csv/level5_test.csv"
-            self.__currentLevelBackground = "sprites/placeholder/level1background.png"
-            self.__keyInput.keymouseleft = False
-            self.__tiles_path = self.get_file_names("sprites/blocks/grassland")
-            print(self.__tiles_path)
-            return "charactermenu"
-        
-        # no input --> reinitialises own loop
-        self.__clock.tick(30)
-        return "levelmenu"
+        imageTop = Image.open(imageTopPath)
+        imageBottom = Image.open(imageBottomPath)
+        imageBottom.paste(imageTop, (0,0))#, mask = __imageTop)
+        imageBottom.save(newImageName) 
     
-
     def get_file_names(self, directory :str):
         file_names = []
         files = os.listdir(directory)
@@ -202,15 +73,167 @@ class Menu():
                 file_names.append(os.path.join(directory, filename))
         return file_names
 
+    def getScreenResoltution(self):
+        return self.__screenResolution
+    
+
+    def getScreen(self):
+        return self.__screen
+    
+    @property
+    def clockTick(self):
+        return self.__clock_tick
+    
+
+
+class MainMenu(Menu):
+    def __init__(self, clock:pygame.time.Clock):
+        super().__init__()
+        self.__keyInput = KeyInput()
+        self.__clock = clock
 
     
-# ===================== character menu ===================== #
+    def drawMainMenu(self) -> None:        
+        pygame.display.set_caption("super main menu")
+        # loads in the objects and draws the main menu
+        self.drawBackground("sprites/placeholder/mainmenu.png")
+        self.__buttonPlayRect = self.drawButton("sprites/placeholder/buttonplay.png", 150, 0)
+        self.__buttonQuitRect = self.drawButton("sprites/placeholder/buttonquit.png", -150, 0)
+
+        pygame.display.flip()
+
+
+    def mainMenuLoop(self)-> bool:
+        """
+        Loop used for the main menu. Returns "levelmenu" or "quit" on specific input, otherwise returns "mainmenu".
+        """
+
+        # frame and input update
+        self.drawMainMenu()
+
+        self.__keyInput.getInput()
+        mousePosition = pygame.mouse.get_pos()
+
+        # input check
+        if self.__keyInput.keymouseleft:
+
+            if self.__buttonPlayRect.collidepoint(mousePosition):
+                self.__keyInput.keymouseleft = False
+                return "levelmenu"
+        
+            if self.__buttonQuitRect.collidepoint(mousePosition):
+                return "quit"
+            
+        if self.__keyInput.keyescape: # thats the RAGEQUIT button :D
+            return "quit"
+
+        # no input --> reinitialises own loop
+        self.__clock.tick(self.clockTick)
+        return "mainmenu"
+
+
+
+class LevelMenu(Menu):
+    def __init__(self,clock:pygame.time.Clock):
+        super().__init__()
+        self.__keyInput = KeyInput()
+        self.__clock = clock
+
+
     
-    def drawCharacterMenu(self):
+    def drawLevelMenu(self)-> None:
+        pygame.display.set_caption("super level menu")
+        # loads in the objects and draws the level menu
+        self.drawBackground("sprites/placeholder/levelmenu.png")
+        self.__level_1_Rect = self.drawButton("sprites/icons/level_icons/level1.png", 250, 500)
+        self.__level_2_Rect = self.drawButton("sprites/icons/level_icons/level2.png", 250, 250)
+        self.__level_3_Rect = self.drawButton("sprites/icons/level_icons/level3.png", 250, 0)
+        self.__level_4_Rect = self.drawButton("sprites/icons/level_icons/level4.png", 250, -250)
+        self.__level_5_Rect = self.drawButton("sprites/icons/level_icons/level5.png", 250, -500)
+
+        self.__tiles_path = []
+
+        pygame.display.flip()
+
+    def levelMenuLoop(self)->bool:
+        """
+        Loop used for the level menu. Returns "levelmenu" or "mainmenu" on specific input, otherwise returns "levelmenu".
+        """
+        self.drawLevelMenu()
+        # frame and input update
+        self.__keyInput.getInput()
+        mousePosition = pygame.mouse.get_pos()
+        
+
+        # input check
+        if self.__keyInput.keyescape:
+            self.__keyInput.keyescape = False
+            return "mainmenu"
+
+        elif self.__keyInput.keymouseleft and self.__level_1_Rect.collidepoint(mousePosition):
+            self.__currentLevel = "sprites/blocks/csv/level1_grassland.csv"
+            self.__currentLevelBackground = "sprites/placeholder/level1background.png"
+            self.__keyInput.keymouseleft = False
+            self.__tiles_path = self.get_file_names("sprites/blocks/grassland")
+            print(self.get_file_names("sprites/blocks/grassland"))
+            return "charactermenu"
+        
+        elif self.__keyInput.keymouseleft and self.__level_2_Rect.collidepoint(mousePosition):
+            self.__currentLevel = "sprites/blocks/csv/level2_desert.csv"
+            self.__currentLevelBackground = "sprites/placeholder/level1background.png"
+            self.__keyInput.keymouseleft = False
+            self.__tiles_path = self.get_file_names("sprites/blocks/desert")
+            print(self.__tiles_path)
+            return "charactermenu"
+        
+        elif self.__keyInput.keymouseleft and self.__level_3_Rect.collidepoint(mousePosition):
+            self.__currentLevel = "sprites/blocks/csv/level3_grassland.csv"
+            self.__currentLevelBackground = "sprites/placeholder/level1background.png"
+            self.__keyInput.keymouseleft = False
+            self.__tiles_path = self.get_file_names("sprites/blocks/grassland_2")
+            print(self.__tiles_path)
+            return "charactermenu"
+        
+        elif self.__keyInput.keymouseleft and self.__level_4_Rect.collidepoint(mousePosition):
+            self.__currentLevel = "sprites/blocks/csv/level4_snowland.csv"
+            self.__currentLevelBackground = "sprites/placeholder/level1background.png"
+            self.__keyInput.keymouseleft = False
+            self.__tiles_path = self.get_file_names("sprites/blocks/snowland")
+            print(self.__tiles_path)
+            return "charactermenu"
+        
+        elif self.__keyInput.keymouseleft and self.__level_5_Rect.collidepoint(mousePosition):
+            self.__currentLevel = "sprites/blocks/csv/level5_test.csv"
+            self.__currentLevelBackground = "sprites/placeholder/level1background.png"
+            self.__keyInput.keymouseleft = False
+            self.__tiles_path = self.get_file_names("sprites/blocks/grassland")
+            print(self.__tiles_path)
+            return "charactermenu"
+        
+        # no input --> reinitialises own loop
+        self.__clock.tick(self.clockTick)
+        return "levelmenu"
+    
+    def getTilesPath(self)->list:
+        return self.__tiles_path
+    
+    def getCurrentLevel(self)->str:
+        return self.__currentLevel
+    
+    def getCurrentLevelBackground(self)->str:
+        return self.__currentLevelBackground
+    
+class CharacterMenu(Menu):
+    def __init__(self, clock:pygame.time.Clock):
         """
         Method to fully initate the creation of the character menu.
         """
+        super().__init__()
+        self.__keyInput = KeyInput()
+        self.__clock = clock
 
+
+    def drawCharacterMenu(self)-> None:
         pygame.display.set_caption("super character menu")
 
         # loads in the objects and draws the level menu
@@ -236,94 +259,114 @@ class Menu():
         self.mergeImage("sprites/characters/poo/po.gif","sprites/icons/character_background.gif","sprites/characters/poo/buttonImage.gif")
         self.__buttonPoRect = self.drawButton("sprites/characters/poo/buttonImage.gif", 250, -500)
         pygame.display.flip()
+        
+    def characterMenuLoop(self)->bool:
+            """
+            Loop used for the character menu. Returns "gameloop" or "levelmenu" on specific input, otherwise returns "charactermenu".
+            """
 
-    def characterMenuLoop(self):
-        """
-        Loop used for the character menu. Returns "gameloop" or "levelmenu" on specific input, otherwise returns "charactermenu".
-        """
+            # frame and input 
+            self.drawCharacterMenu()
 
-        # frame and input update
-        self.drawCharacterMenu()
-        self.__keyInput.getInput()
-        self.__mousePosition = pygame.mouse.get_pos()
+            self.__keyInput.getInput()
+            mousePosition = pygame.mouse.get_pos()
 
-        # input check
-        if self.__keyInput.keyescape:
-            self.__keyInput.keyescape = False
-            return "levelmenu"
-        
-        if self.__keyInput.keymouseleft and self.__buttonWuRect.collidepoint(self.__mousePosition):
-            self.__currentCharacterSkin = "sprites/characters/wu/wu_image.gif"
-            self.__death_path =  "sprites/characters/wu/wu_death.gif"
-            self.__jump_path = "sprites/characters/wu/wu_jump.gif"
-            self.__keyInput.keymouseleft = False
-            return "gameloop"
-        
-        if self.__keyInput.keymouseleft and self.__buttonFichRect.collidepoint(self.__mousePosition):
-            self.__currentCharacterSkin = "sprites/characters/fich/fich_image.gif"
-            self.__death_path =  "sprites/characters/fich/fich_death.gif"
-            self.__jump_path = "sprites/characters/fich/fich_jump.gif"
-            self.__keyInput.keymouseleft = False
-            return "gameloop"
-        
-        if self.__keyInput.keymouseleft and self.__buttonAmogusRect.collidepoint(self.__mousePosition):
-            self.__currentCharacterSkin = "sprites/characters/amogus/amogus_image.gif"
-            self.__death_path =  "sprites/characters/amogus/amogus_death.gif"
-            self.__jump_path = "sprites/characters/amogus/amogus_jump.gif"
-            self.__keyInput.keymouseleft = False
-            return "gameloop"
-        
-        if self.__keyInput.keymouseleft and self.__buttonPacmanRect.collidepoint(self.__mousePosition):
-            self.__currentCharacterSkin = "sprites/characters/pacman/pacman_image.gif"
-            self.__death_path =  "sprites/characters/pacman/pacman_death.gif"
-            self.__jump_path = "sprites/characters/pacman/pacman_jump.gif"
-            self.__keyInput.keymouseleft = False
-            return "gameloop"
-        
-        if self.__keyInput.keymouseleft and self.__buttonPoRect.collidepoint(self.__mousePosition):
-            self.__currentCharacterSkin = "sprites/characters/poo/po_image.gif"
-            self.__death_path =  "sprites/characters/poo/po_death.gif"
-            self.__jump_path = "sprites/characters/poo/po_jump.gif"
-            self.__keyInput.keymouseleft = False
-            return "gameloop"
-        
-        # no input --> reinitialises own loop
-        self.__clock.tick(30)
-        return "charactermenu"
+            # input check
+            if self.__keyInput.keyescape:
+                self.__keyInput.keyescape = False
+                return "levelmenu"
+            
+            if self.__keyInput.keymouseleft and self.__buttonWuRect.collidepoint(mousePosition):
+                self.__currentCharacterSkin = "sprites/characters/wu/wu_image.gif"
+                self.__death_path =  "sprites/characters/wu/wu_death.gif"
+                self.__jump_path = "sprites/characters/wu/wu_jump.gif"
+                self.__keyInput.keymouseleft = False
+                return "gameloop"
+            
+            if self.__keyInput.keymouseleft and self.__buttonFichRect.collidepoint(mousePosition):
+                self.__currentCharacterSkin = "sprites/characters/fich/fich_image.gif"
+                self.__death_path =  "sprites/characters/fich/fich_death.gif"
+                self.__jump_path = "sprites/characters/fich/fich_jump.gif"
+                self.__keyInput.keymouseleft = False
+                return "gameloop"
+            
+            if self.__keyInput.keymouseleft and self.__buttonAmogusRect.collidepoint(mousePosition):
+                self.__currentCharacterSkin = "sprites/characters/amogus/amogus_image.gif"
+                self.__death_path =  "sprites/characters/amogus/amogus_death.gif"
+                self.__jump_path = "sprites/characters/amogus/amogus_jump.gif"
+                self.__keyInput.keymouseleft = False
+                return "gameloop"
+            
+            if self.__keyInput.keymouseleft and self.__buttonPacmanRect.collidepoint(mousePosition):
+                self.__currentCharacterSkin = "sprites/characters/pacman/pacman_image.gif"
+                self.__death_path =  "sprites/characters/pacman/pacman_death.gif"
+                self.__jump_path = "sprites/characters/pacman/pacman_jump.gif"
+                self.__keyInput.keymouseleft = False
+                return "gameloop"
+            
+            if self.__keyInput.keymouseleft and self.__buttonPoRect.collidepoint(mousePosition):
+                self.__currentCharacterSkin = "sprites/characters/poo/po_image.gif"
+                self.__death_path =  "sprites/characters/poo/po_death.gif"
+                self.__jump_path = "sprites/characters/poo/po_jump.gif"
+                self.__keyInput.keymouseleft = False
+                return "gameloop"
+            
+            # no input --> reinitialises own loop
+            self.__clock.tick(self.clockTick)
+            return "charactermenu"
 
-# ===================== win menu ===================== #
-    def drawWinMenu(self):
+    def getCurrentCharacterSkin(self)->str:
+        return self.__currentCharacterSkin
+    
+    def getJumpPath(self)->str:
+        return self.__jump_path
+
+    def getDeathPath(self)->str:
+        return self.__death_path
+
+class WinMenu(Menu):
+    def __init__(self):
         """
         draws the win menu or show it on the screen
         """
+        super().__init__()
+        self.__keyInput = KeyInput()
+    
+    def drawWinMenu(self)-> None:
         pygame.display.set_caption("you won supa nite fort")
 
         # loads in the objects and draws the level menu
         self.drawBackground("sprites/placeholder/level1background.png")
 
         self.__menuButton = self.drawButton("sprites/placeholder/menu.png", 0, 0)
-        self.__winLabel = self.drawButton("sprites/placeholder/YouWin.png", 250, 0) 
-
+        self.drawButton("sprites/placeholder/YouWin.png", 250, 0)
 
     def winMenuLoop(self):
         """
         Loop used for the menu if you win the game
         """
-        self.drawWinMenu()
         self.__keyInput.getInput()
-        self.__mousePosition = pygame.mouse.get_pos()
+        mousePosition = pygame.mouse.get_pos()
 
-        if self.__keyInput.keymouseleft and self.__menuButton.collidepoint(self.__mousePosition):
+        if self.__keyInput.keymouseleft and self.__menuButton.collidepoint(mousePosition):
             self.__keyInput.keymouseleft = False
             return "mainmenu"
         
         pygame.display.flip()
         
-        return "winmenu"
+        return "winmenu" 
+    
 
-
-        
-# ===================== game menu ===================== #
+class GameLoop():
+    def __init__(self, screenResolution:pygame.math.Vector2, currentLevel:str, currentLevelBackground:str, currentCharacterSkin:str, death_path:str, jump_path:str, tiles_path:list, screen:pygame.surface.Surface) -> None:
+        self.__screenResolution = screenResolution
+        self.__currentLevel = currentLevel
+        self.__currentLevelBackground = currentLevelBackground
+        self.__currentCharacterSkin = currentCharacterSkin
+        self.__death_path = death_path
+        self.__jump_path = jump_path
+        self.__tiles_path = tiles_path
+        self.__screen = screen
 
     def gameLoop(self):
         """
@@ -345,11 +388,4 @@ class Menu():
         if self.__mainLoop.getWin():
             return "winmenu"
 
-        return "mainmenu" #TODO replace mainmenu with pausemenu in this instance
-    
-# ===================== pause menu ===================== #
-                
-    def drawPauseMenu(self):
-        pass
-    # TODO sets the background to a blurred or darker current game frame
-    # TODO button to resume, quit to main menu and to quit the game
+        return "mainmenu"
